% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GCM}
\alias{GCM}
\alias{gcm_pr_ch}
\alias{gcm_tas_ch}
\title{Data z globálního klimatického modelu}
\format{
RasterBrick

An object of class \code{RasterBrick} of dimension 37 x 43 x 12.

An object of class \code{RasterBrick} of dimension 37 x 43 x 12.
}
\usage{
gcm_pr_ch

gcm_tas_ch
}
\description{
Změny srážek a teploty pro jednotlivé měsíce dle simulace CMIP5 model HadGEM-ESM2 podle RCP8.5 mezi obdobími 2070-2100 a 1970-2000
}
\keyword{datasets}
