% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hydrometeo}
\alias{hydrometeo}
\title{Hydrometeorologická data pro 28 povodí v ČR}
\format{
data.table obsahující následující proměnné

\describe{
\item{DBCN}{databankové číslo (identifikátor povodí) - pomocí něj je možné propojit s datasetem \link{povodi}}
\item{AREA}{plocha povodí v m2}
\item{DTM}{datum}
\item{Q}{průtok [m3/s]}
\item{P}{denní srážky [mm]}
\item{T}{denní teplota [st. C]}
\item{R}{denní odtok [mm]}
}
}
\usage{
hydrometeo
}
\description{
Hydrometeorologická data pro 28 povodí v ČR
}
\keyword{datasets}
